reorganise lib
