+++ example/example.dart
@@ -71,37 +71,40 @@ void main() {
   ));*/
 
   ///http://www.chartjs.org/docs/#line-chart-dataset-structure
-  /* Chart.Line(ctx, new LineChartConfiguration(
+  Chart.Line(ctx, new LineChartConfiguration(
       data: new ChartData(
           labels: ['Item 1', 'Item 2', 'Item 3'],
           datasets: [
             new LineChartDataSet(
                 label: "My First dataset",
                 fill: false,
-                lineTension: 0.1,
-                backgroundColor: ["rgba(75,192,192,0.4)"],
+                lineTension: 0.2,
+                backgroundColor: ["rgba(75,192,192,0)"],
                 borderColor: ["rgba(75,192,192,1)"],
                 borderCapStyle: 'butt',
                 borderDash: [],
                 borderDashOffset: 0.0,
                 borderJoinStyle: 'miter',
-                pointBorderColor: "rgba(75,192,192,1)",
-                pointBackgroundColor: "#fff",
+                pointBorderColor: "rgba(75,192,192,0)",
+                pointBackgroundColor: "#ff0000",
                 pointBorderWidth: [1],
                 pointHoverRadius: [5],
                 pointHoverBackgroundColor: ["rgba(75,192,192,1)"],
                 pointHoverBorderColor: ["rgba(220,220,220,1)"],
                 pointHoverBorderWidth: [2],
-                pointRadius: [1],
+                pointRadius: [5],
                 pointHitRadius: [10],
                 data: [65, 59, 80, 81, 56, 55, 40],
-                spanGaps: false,)
+                spanGaps: false,
+                showLines: false)
           ]
-      )
-  ));*/
+      ),
+      options: new ChartSettings(scales: new Scales(xAxes: [ new Axis(display: false)]))
+  )
+  );
 
   ///http://www.chartjs.org/docs/#line-chart-scatter-line-charts
-  /*Chart.Line(ctx, new LineChartConfiguration(
+/*Chart.Line(ctx, new LineChartConfiguration(
       data: new ChartData(
           datasets: [new LineChartDataSet(
               label: 'Scatter Dataset',
@@ -114,7 +117,7 @@ void main() {
   ));*/
 
   ///http://www.chartjs.org/docs/#line-chart-stacked-charts
-  Chart.Line(ctx, new LineChartConfiguration(
+/*Chart.Line(ctx, new LineChartConfiguration(
       data: new ChartData(
           labels: ['Item 1', 'Item 2', 'Item 3'],
           datasets: [
@@ -129,5 +132,5 @@ void main() {
       options: new ChartSettings(
           scales: new Scales(yAxes: [ new Axis(stacked: true)])
       )
-  ));
+  ));*/
 }
